+++ src/pb-image-row/view.js
@@ -1,4 +1,0 @@
-/**
- * PB Image Row Block
- * View JS
- **/