+++ src/pb-image-stack/view.js
@@ -1,4 +1,0 @@
-/**
- * PB Image Stack Block
- * View JS
- **/